+++ src/ADN.Extensions/Models/DateTimeExtensions.cs
@@ -12,11 +12,11 @@ namespace ADN.Extensions
         /// <summary>
         /// Converts DateTime to unix time stamp.
         /// </summary>
-        /// <param name="date">The DateTime.</param>
+        /// <param name="dateTime">The DateTime.</param>
         /// <returns>The unix time stamp.</returns>
         /// <example>
         /// <code lang="csharp">
-        /// DateTime date = DateTime.Parse("02/16/2008 12:15:12", new CultureInfo("en-US"));
+        /// var dateTime = DateTime.Parse("02/16/2008 12:15:12", new CultureInfo("en-US"));
         /// var result = date.DateTimeToUnixTimeStamp();
         /// 
         /// /*
@@ -24,9 +24,9 @@ namespace ADN.Extensions
         /// */
         /// </code>
         /// </example>
-        public static long DateTimeToUnixTimeStamp(this DateTime date)
+        public static long DateTimeToUnixTimeStamp(this DateTime dateTime)
         {
-            long unixTimestamp = date.Ticks - new DateTime(1970, 1, 1).Ticks;
+            long unixTimestamp = dateTime.Ticks - new DateTime(1970, 1, 1).Ticks;
             unixTimestamp /= TimeSpan.TicksPerSecond;
 
             return unixTimestamp;
@@ -49,10 +49,55 @@ namespace ADN.Extensions
         /// </example>
         public static DateTime UnixTimeStampToDateTime(this long unixTimestamp)
         {
-            DateTime date = new DateTime(1970, 1, 1);
+            var dateTime = new DateTime(1970, 1, 1);
             long unixTimeStampInTicks = unixTimestamp * TimeSpan.TicksPerSecond;
 
-            return new DateTime(date.Ticks + unixTimeStampInTicks);
+            return new DateTime(dateTime.Ticks + unixTimeStampInTicks);
+        }
+
+        /// <summary>
+        /// Returns the DateTime with the ticks trimmed.
+        /// </summary>
+        /// <param name="dateTime">The DateTime.</param>
+        /// <returns>A DateTime with the ticks trimmed.</returns>
+        /// <example>
+        /// <code lang="csharp">
+        /// var dateTime = DateTime.Now.TrimTicks();
+        /// </code>
+        /// </example>
+        public static DateTime TrimTicks(this DateTime dateTime)
+        {
+            return new DateTime(dateTime.Year, dateTime.Month, dateTime.Day, dateTime.Hour, dateTime.Minute, dateTime.Second, dateTime.Millisecond, dateTime.Kind);
+        }
+
+        /// <summary>
+        /// Returns the DateTime with the milliseconds trimmed.
+        /// </summary>
+        /// <param name="dateTime">The DateTime.</param>
+        /// <returns>A DateTime with the milliseconds trimmed.</returns>
+        /// <example>
+        /// <code lang="csharp">
+        /// var dateTime = DateTime.Now.TrimMilliseconds();
+        /// </code>
+        /// </example>
+        public static DateTime TrimMilliseconds(this DateTime dateTime)
+        {
+            return new DateTime(dateTime.Year, dateTime.Month, dateTime.Day, dateTime.Hour, dateTime.Minute, dateTime.Second, dateTime.Kind);
+        }
+
+        /// <summary>
+        /// Returns a Datetime with the seconds trimmed.
+        /// </summary>
+        /// <param name="dateTime">The DateTime.</param>
+        /// <returns>A DateTime with the seconds trimmed.</returns>
+        /// <example>
+        /// <code lang="csharp">
+        /// var dateTime = DateTime.Now.TrimSecondsAndMilliseconds();
+        /// </code>
+        /// </example>
+        public static DateTime TrimSecondsAndMilliseconds(this DateTime dateTime)
+        {
+            return new DateTime(dateTime.Year, dateTime.Month, dateTime.Day, dateTime.Hour, dateTime.Minute, 0, 0, dateTime.Kind);
         }
     }
 }